+++ common/src/main/java/io/github/flemmli97/runecraftory/common/attachment/player/Party.java
@@ -1,0 +1,4 @@
+package io.github.flemmli97.runecraftory.common.attachment.player;
+
+public class Party {
+}