+++ common/src/main/java/io/github/flemmli97/runecraftory/common/blocks/BlockCashRegister.java
@@ -1,0 +1,42 @@
+package io.github.flemmli97.runecraftory.common.blocks;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.HorizontalDirectionalBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+
+import javax.annotation.Nullable;
+
+public class BlockCashRegister extends HorizontalDirectionalBlock {
+
+    public static final VoxelShape SHAPE = Block.box(0, 0, 0, 16, 12, 16);
+
+    /**
+     * protected super constructor...
+     */
+    public BlockCashRegister(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    @Nullable
+    public BlockState getStateForPlacement(BlockPlaceContext ctx) {
+        return this.defaultBlockState().setValue(FACING, ctx.getHorizontalDirection());
+
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(FACING);
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return SHAPE;
+    }
+}