+++ common/src/main/java/io/github/flemmli97/runecraftory/common/config/values/EntityProperties.java
@@ -4,12 +4,10 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import io.github.flemmli97.runecraftory.common.lib.LibAttributes;
 import io.github.flemmli97.runecraftory.common.registry.ModAttributes;
-import io.github.flemmli97.tenshilib.api.config.SimpleItemStackWrapper;
 import io.github.flemmli97.tenshilib.platform.PlatformUtils;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.ai.attributes.Attribute;
 
-import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
@@ -32,8 +30,10 @@ public class EntityProperties {
     private final boolean flying;
     private Map<Attribute, Double> baseValues;
     private Map<Attribute, Double> levelGains;
+    private final int size;
+    private final boolean needsRoof;
 
-    private EntityProperties(List<String> baseValues, List<String> gains, int xp, int money, float tamingChance, boolean ridable, boolean flying) {
+    private EntityProperties(List<String> baseValues, List<String> gains, int xp, int money, float tamingChance, boolean ridable, boolean flying, int size, boolean needsRoof) {
         this.confAttributes = baseValues;
         this.confGains = gains;
         this.xp = xp;
@@ -41,13 +41,8 @@ public class EntityProperties {
         this.taming = tamingChance;
         this.ridable = ridable;
         this.flying = flying;
-    }
-
-    public static List<String> dailyToString(Map<SimpleItemStackWrapper, Integer> map) {
-        List<String> list = new ArrayList<>();
-        for (Map.Entry<SimpleItemStackWrapper, Integer> e : map.entrySet())
-            list.add(e.getKey().writeToString() + "-" + e.getValue());
-        return list;
+        this.size = size;
+        this.needsRoof = needsRoof;
     }
 
     public List<String> attString() {
@@ -104,6 +99,14 @@ public class EntityProperties {
         return this.taming;
     }
 
+    public int getSize() {
+        return this.size;
+    }
+
+    public boolean needsRoof() {
+        return this.needsRoof;
+    }
+
     public static class Builder {
 
         private final Set<String> baseValues = new LinkedHashSet<>();
@@ -113,6 +116,8 @@ public class EntityProperties {
         private float taming;
         private boolean ridable;
         private boolean flying;
+        private int size = 1;
+        private boolean needsRoof = true;
 
         public Builder putAttributes(String att, double val) {
             this.baseValues.add(att + "-" + val);
@@ -157,8 +162,18 @@ public class EntityProperties {
             return this;
         }
 
+        public Builder setBarnOccupancy(int size) {
+            this.size = Math.max(1, size);
+            return this;
+        }
+
+        public Builder doesntNeedBarnRoof() {
+            this.needsRoof = false;
+            return this;
+        }
+
         public EntityProperties build() {
-            return new EntityProperties(Lists.newArrayList(this.baseValues), Lists.newArrayList(this.gains), this.xp, this.money, this.taming, this.ridable, this.flying);
+            return new EntityProperties(Lists.newArrayList(this.baseValues), Lists.newArrayList(this.gains), this.xp, this.money, this.taming, this.ridable, this.flying, this.size, this.needsRoof);
         }
     }
 }