+++ common/src/main/java/io/github/flemmli97/runecraftory/common/entities/npc/EnumShop.java
@@ -3,6 +3,8 @@ package io.github.flemmli97.runecraftory.common.entities.npc;
 import io.github.flemmli97.runecraftory.common.registry.ModPoiTypes;
 import net.minecraft.world.entity.ai.village.poi.PoiType;
 
+import javax.annotation.Nullable;
+import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 public enum EnumShop {
@@ -13,16 +15,23 @@ public enum EnumShop {
     WEAPON(() -> PoiType.TOOLSMITH, "shop_weapon"),
     CLINIC(() -> PoiType.CLERIC, "shop_clinic"),
     FOOD(() -> PoiType.BUTCHER, "shop_food"),
-    MAGIC(ModPoiTypes.skills, "shop_magic"),
-    RUNESKILL(ModPoiTypes.skills, "shop_runeskill"),
+    MAGIC(ModPoiTypes.CASH_REGISTER, "shop_magic"),
+    RUNESKILL(ModPoiTypes.CASH_REGISTER, "shop_runeskill"),
     RANDOM(() -> null, "shop_random");
 
     public final Supplier<PoiType> poiType;
+
+    @Nullable
+    public final Predicate<PoiType> predicate;
 
     public final String translationKey;
 
     EnumShop(Supplier<PoiType> poiType, String translationKey) {
         this.poiType = poiType;
         this.translationKey = translationKey;
+        if (poiType == ModPoiTypes.CASH_REGISTER)
+            this.predicate = t -> this.poiType.get().getPredicate().test(t);
+        else
+            this.predicate = this.poiType == null ? null : t -> this.poiType.get().getPredicate().test(t) || ModPoiTypes.CASH_REGISTER.get().getPredicate().test(t);
     }
 }