+++ common/src/main/java/io/github/flemmli97/runecraftory/common/registry/ModTags.java
@@ -132,6 +132,8 @@ public class ModTags {
 
     public static final TagKey<Block> MONSTER_CLEARABLE = blockMod("monster_clearable");
 
+    public static final TagKey<Block> BARN_GROUND = blockMod("barn_ground");
+
     //Entities
     public static final TagKey<EntityType<?>> MONSTERS = TagKey.create(Registry.ENTITY_TYPE_REGISTRY, new ResourceLocation(RuneCraftory.MODID, "monsters"));
     public static final TagKey<EntityType<?>> BOSS_MONSTERS = TagKey.create(Registry.ENTITY_TYPE_REGISTRY, new ResourceLocation(RuneCraftory.MODID, "boss_monsters"));