+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/config/values/EntityPropertySpecs.java
@@ -2,6 +2,7 @@ package io.github.flemmli97.runecraftory.forge.config.values;
 
 import io.github.flemmli97.runecraftory.common.config.values.EntityProperties;
 import io.github.flemmli97.tenshilib.api.config.SimpleItemStackWrapper;
+import io.github.flemmli97.tenshilib.common.config.CommentedJsonConfig;
 import net.minecraftforge.common.ForgeConfigSpec;
 
 import java.util.ArrayList;
@@ -16,6 +17,8 @@ public class EntityPropertySpecs {
     public final ForgeConfigSpec.DoubleValue taming;
     public final ForgeConfigSpec.BooleanValue ridable;
     public final ForgeConfigSpec.BooleanValue flying;
+    public final ForgeConfigSpec.IntValue size;
+    public final ForgeConfigSpec.BooleanValue needsRoof;
 
     public EntityPropertySpecs(ForgeConfigSpec.Builder builder, EntityProperties def) {
         this.baseValues = builder.comment("Base Values at level 1 of the mob").define("Base Values", def.attString());
@@ -25,6 +28,8 @@ public class EntityPropertySpecs {
         this.taming = builder.comment("Base chance to tame this mob").defineInRange("Taming Chance", def.tamingChance(), 0, 1);
         this.ridable = builder.define("Ridable", def.ridable());
         this.flying = builder.define("Can Fly", def.flying());
+        this.size = builder.comment("The amount of space this monster takes in a barn").defineInRange("Size", def.getSize(), 1, Integer.MAX_VALUE);
+        this.needsRoof = builder.define("If the monsters barn needs a roof", def.needsRoof());
     }
 
     public static EntityProperties ofSpec(EntityPropertySpecs spec) {
@@ -50,6 +55,9 @@ public class EntityPropertySpecs {
             builder.setRidable();
         if (spec.flying.get())
             builder.setFlying();
+        builder.setBarnOccupancy(spec.size.get());
+        if(!spec.needsRoof.get())
+            builder.doesntNeedBarnRoof();
         return builder.build();
     }