+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/data/BlockTagGen.java
@@ -74,10 +74,13 @@ public class BlockTagGen extends BlockTagsProvider {
         }
 
         this.tag(BlockTags.MINEABLE_WITH_PICKAXE).add(ModBlocks.forge.get(), ModBlocks.cooking.get(), ModBlocks.chemistry.get(), ModBlocks.accessory.get(),
-                ModBlocks.bossSpawner.get());
+                ModBlocks.bossSpawner.get(), ModBlocks.cashRegister.get());
         this.tag(BlockTags.MINEABLE_WITH_SHOVEL).add(ModBlocks.farmland.get());
         this.tag(BlockTags.MINEABLE_WITH_AXE).add(ModBlocks.board.get(), ModBlocks.shipping.get());
+        this.tag(BlockTags.MINEABLE_WITH_HOE).add(ModBlocks.monsterBarn.get());
 
         this.tag(ModTags.FARMLAND).add(Blocks.FARMLAND, ModBlocks.farmland.get());
+
+        this.tag(ModTags.BARN_GROUND).add(Blocks.HAY_BLOCK);
     }
 }